refactor(admin): move search endpoints to /search and lists to canonical paths
