+++ apps/core/src/routes/v1/admin/organizations/[slug]/get.ts
@@ -1,13 +1,13 @@
 import { createRoute } from "@hono/zod-openapi";
-import { organizationRepository } from "@sokosumi/database/repositories";
 
+import { buildAdminOrganizationOverviewDetail } from "@/helpers/admin-organization-overview.js";
 import { notFound } from "@/helpers/error";
 import { jsonErrorResponse, jsonSuccessResponse } from "@/helpers/openapi";
 import { ok } from "@/helpers/response";
 import prisma from "@/lib/db/prisma";
 import type { OpenAPIHonoWithAuth } from "@/lib/hono";
 import {
-  adminOrganizationOptionSchema,
+  adminOrganizationOverviewDetailSchema,
   adminOrganizationSlugParamSchema,
 } from "@/schemas/admin.schema";
 
@@ -16,22 +16,15 @@ const route = createRoute({
   path: "/{slug}",
   operationId: "getAdminOrganizationBySlug",
   description:
-    "Fetch a single organization's limited info by slug (admin only). Returns 404 when no organization matches.",
+    "Full organization overview with billing, subscription, seats, and credits (admin only).",
   tags: ["Admin"],
   request: {
     params: adminOrganizationSlugParamSchema,
   },
   responses: {
     200: jsonSuccessResponse(
-      adminOrganizationOptionSchema,
-      "Organization matching the slug",
-      {
-        data: { id: "org_123", name: "Acme Corp", slug: "acme-corp" },
-        meta: {
-          timestamp: "2025-01-01T00:00:00.000Z",
-          requestId: "550e8400-e29b-41d4-a716-446655440000",
-        },
-      },
+      adminOrganizationOverviewDetailSchema,
+      "Organization overview for the admin console",
     ),
     401: jsonErrorResponse("Unauthorized"),
     403: jsonErrorResponse("Forbidden"),
@@ -43,23 +36,12 @@ export default function mount(app: OpenAPIHonoWithAuth) {
   app.openapi(route, async (c) => {
     const { slug } = c.req.valid("param");
 
-    const organization =
-      await organizationRepository.getOrganizationLimitedInfoBySlug(
-        slug,
-        prisma,
-      );
+    const detail = await buildAdminOrganizationOverviewDetail(slug, prisma);
 
-    if (!organization) {
+    if (!detail) {
       throw notFound("Organization not found");
     }
 
-    return ok(
-      c,
-      adminOrganizationOptionSchema.parse({
-        id: organization.id,
-        name: organization.name,
-        slug: organization.slug,
-      }),
-    );
+    return ok(c, adminOrganizationOverviewDetailSchema.parse(detail));
   });
 }
